Remove MutableConfigService, update test cases
