+++ packages/config/src/ConfigService/index.ts
@@ -1,3 +1,2 @@
 export * from './ConfigService';
 export * from './ConfigService.interfaces';
-export * from './MutableConfigService';
